requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0

# Data & ML
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0

# LLM
groq>=0.11.0

# UI
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.18.0

# Phase 2/3 (install now, use later)
reportlab>=4.0.0

# Dev
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.0
